Rename package from duke to bob
