+++ docs/README.md
@@ -22,9 +22,9 @@ Duke provides a command-line tool for managing your tasks. It helps you track th
 1. Ensure you have Java `11` or above installed in your computer. 
 2. Download the latest `ip.jar` from [here](https://github.com/lctxct/ip/releases/tag/A-Release).
 3. Copy the file to the folder you want to use as the *home folder* for Duke. 
-4. Open a command terminal, locate the path to the folder you put the jar file in, and `cd` (change directory) into that folder. For example, if your file path is `C:\Users\potato\duke\ip.jar`, the full command will be 
+4. Open a command terminal, locate the path to the folder you put the jar file in, and `cd` (change directory) into that folder. For example, if your file path is `C:\Users\potato\bob\ip.jar`, the full command will be 
     ```
-    cd C:\Users\potato\duke\ip.jar
+    cd C:\Users\potato\bob\ip.jar
     ```
 5. Once you're in the directory, use the following command to run the application.
     ```
@@ -43,15 +43,15 @@ Duke provides a command-line tool for managing your tasks. It helps you track th
     What can I do for you?
     ____________________________________________________________
     ```
-    If you encounter the error message "Unable to find file. Defaulting to empty list...", this means that Duke is unable to find a save file at the default path `./data/duke.txt`. To remedy this, you may first type `bye` and press Enter to exit the program, and then execute the following. 
+    If you encounter the error message "Unable to find file. Defaulting to empty list...", this means that Duke is unable to find a save file at the default path `./data/bob.txt`. To remedy this, you may first type `bye` and press Enter to exit the program, and then execute the following. 
     
     For **Windows**: 
     ```
-    New-Item -ItemType Directory -Path .\data; New-Item -ItemType File -Path .\data\duke.txt
+    New-Item -ItemType Directory -Path .\data; New-Item -ItemType File -Path .\data\bob.txt
     ```
     For **Unix**: 
     ```
-    mkdir -p ./data && touch ./data/duke.txt
+    mkdir -p ./data && touch ./data/bob.txt
     ```
     Restart the program with the same `java -jar ip.jar` command. Note that the error message should no be present. 
 6. Now that the application is running, here are some sample commands you can try: 
@@ -121,10 +121,10 @@ Format: `delete <index>`
 Example: `delete 1`
 
 ### Exiting the program: `bye`
-Exits the program. If the save file `./data/duke.txt` exists, contents of the task list will be written into the save file. View [Saving the data](#saving-the-data) for more information on creating the save file. 
+Exits the program. If the save file `./data/bob.txt` exists, contents of the task list will be written into the save file. View [Saving the data](#saving-the-data) for more information on creating the save file. 
 
 ### Saving the data
-If the save file `.\data\duke.txt` exists, items remaining in the task list will be saved in the save file when the app exits, and will be reloaded the next time Duke is launched. 
+If the save file `.\data\bob.txt` exists, items remaining in the task list will be saved in the save file when the app exits, and will be reloaded the next time Duke is launched. 
 
 If the save file cannot be found, upon starting up Duke will prepend to the welcome message the following error: "Unable to find file. Defaulting to empty list...". 
 
@@ -132,11 +132,11 @@ To remedy this, you may first type `bye` and press Enter to exit the program, an
     
 - For **Windows**: 
     ```
-    New-Item -ItemType Directory -Path .\data; New-Item -ItemType File -Path .\data\duke.txt
+    New-Item -ItemType Directory -Path .\data; New-Item -ItemType File -Path .\data\bob.txt
     ```
 - For **Unix**: 
     ```
-    mkdir -p ./data && touch ./data/duke.txt
+    mkdir -p ./data && touch ./data/bob.txt
     ```
 Note that when the program is restarted with the `java -jar ip.jar` command, the error message should no longer be present.
