+++ src/bob/DukeException.java
@@ -1,0 +1,7 @@
+package bob;
+
+public class DukeException extends Exception {
+    public DukeException (String str) {
+        super(str);
+    }
+}
