+++ src/bob/tasklist/TaskList.java
@@ -1,10 +1,10 @@
-package duke.tasklist;
+package bob.tasklist;
 
-import duke.DukeException;
-import duke.deadline.Deadline;
-import duke.event.Event;
-import duke.task.Task;
-import duke.todo.Todo;
+import bob.DukeException;
+import bob.deadline.Deadline;
+import bob.event.Event;
+import bob.task.Task;
+import bob.todo.Todo;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
