--- src/duke/todo/Todo.java
@@ -1,21 +1,0 @@
-package duke.todo;
-
-import duke.task.Task;
-
-public class Todo extends Task {
-
-    /**
-     * Constructs a new Todo.
-     *
-     * @param description contains description of Todo.
-     */
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String getTypeIcon() {
-        return "T";
-    }
-
-}
