+++ text-ui-test/runtest.bat
@@ -7,7 +7,7 @@ REM delete output from previous run
 if exist ACTUAL.TXT del ACTUAL.TXT
 
 REM compile the code into the bin folder
-javac  -cp ..\src\duke -Xlint:none -d ..\bin ..\src\duke\*.java
+javac  -cp ..\src\bob -Xlint:none -d ..\bin ..\src\bob\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
